+++ packages/stream_video/lib/src/stream_video.dart
@@ -46,6 +46,7 @@ import 'models/call_preferences.dart';
 import 'models/call_received_data.dart';
 import 'models/call_ringing_data.dart';
 import 'models/call_status.dart';
+import 'models/disconnect_reason.dart';
 import 'models/guest_created_data.dart';
 import 'models/push_device.dart';
 import 'models/push_provider.dart';
@@ -997,7 +998,9 @@ class StreamVideo extends Disposable {
     final incomingCall = _state.incomingCall.valueOrNull;
 
     if (activeCall?.callCid.value == cid) {
-      final result = await activeCall?.leave();
+      final result = await activeCall?.leave(
+        reason: DisconnectReason.callEnded(),
+      );
 
       if (result is Failure) {
         _logger.d(() => '[onCallEnded] error leaving call: ${result.error}');
@@ -1006,7 +1009,7 @@ class StreamVideo extends Disposable {
       final status = incomingCall?.state.value.status;
       if (status is CallStatusIncoming && !status.acceptedByMe) {
         final result = await incomingCall?.reject(
-          reason: CallRejectReason.decline(),
+          reason: CallRejectReason.callEnded(),
         );
         if (result is Failure) {
           _logger.d(