+++ packages/stream_video_push_notification/lib/src/stream_video_push_notification.dart
@@ -63,6 +63,7 @@ class StreamVideoPushNotificationManager implements PushNotificationManager {
           _logger.d(
             () => '[subscribeToEvents] Call ended event: ${event.callCid}',
           );
+          RingingEventBroadcaster().suppressEvent();
           endCallByCid(event.callCid.toString());
         }),
       );
@@ -88,6 +89,7 @@ class StreamVideoPushNotificationManager implements PushNotificationManager {
                   '[subscribeToEvents] No participants left, ending call: ${event.callCid}',
             );
 
+            RingingEventBroadcaster().suppressEvent();
             await endCallByCid(event.callCid.toString());
           }
         }),
@@ -107,6 +109,7 @@ class StreamVideoPushNotificationManager implements PushNotificationManager {
                   '[subscribeToEvents] Call rejected by the current user or call owner, ending call: ${event.callCid}',
             );
 
+            RingingEventBroadcaster().suppressEvent();
             await endCallByCid(event.callCid.toString());
           }
         }),